Correcting type to uint8

+++ ConvertLabels.py
@@ -16,8 +16,8 @@ else:
     
     tempdata=readdata
     for i in xrange(0,len(new_list)):
-        dc = readdata==int(old_list[i])
-        tempdata[dc] = int(new_list[i])
+        dc = readdata==np.uint8(old_list[i])
+        tempdata[dc] = np.uint8(new_list[i])
         print old_list[i], '->', new_list[i]
